Move gatsby-node back inline and fix develop mode
