+++ package.json
@@ -109,7 +109,8 @@
     "stylelint-config-prettier": "^8.0.1",
     "stylelint-prettier": "^1.1.2",
     "ts-jest": "^25.2.1",
-    "ts-node": "^8.6.2"
+    "ts-node": "^8.6.2",
+    "tsconfig-paths-webpack-plugin": "^3.2.0"
   },
   "homepage": "https://unicorn-utterances.com",
   "keywords": [
